jsonschema: test cleanup and some new tests
Clenaup:
- Renamed some sub-tests / sub-benchmarks for increased consistency
- Renamed/moved definitions of two "nested" schemas to testutil_test.go
- Moves test utility types, vars and functions to testutil_test.go
- Moves complex schema functions to benchmark_test.go

New tests:
- Object gets dedicated test file with a new test
- New test/benchmark for simpleSchema
- New benchmark for nestedObjects and simpleSchema
